+++ product_code/report_service/impl/src/main/scala/de/upb/cs/uc4/report/impl/actor/TextReport.scala
@@ -3,6 +3,7 @@ package de.upb.cs.uc4.report.impl.actor
 import de.upb.cs.uc4.admission.model.CourseAdmission
 import de.upb.cs.uc4.certificate.model.EncryptedPrivateKey
 import de.upb.cs.uc4.course.model.Course
+import de.upb.cs.uc4.hyperledger.api.model.operation.OperationData
 import de.upb.cs.uc4.matriculation.model.ImmatriculationData
 import de.upb.cs.uc4.user.model.user.User
 import play.api.libs.json.{ Format, Json }
@@ -15,6 +16,8 @@ case class TextReport(
     immatriculationData: Option[ImmatriculationData],
     courses: Option[Seq[Course]],
     admissions: Option[Seq[CourseAdmission]],
+    operations: Seq[OperationData],
+    watchlist: Seq[String],
     timestamp: String
 )