+++ product_code/report_service/CHANGELOG.md
@@ -1,3 +1,8 @@
+# [WIP - v0.17.1](https://github.com/upb-uc4/University-Credits-4.0/compare/report-v0.16.1...report-v0.17.1) (2021-XX-XX)
+## Feature
+## Refactor
+## Bugfix
+
 # [v0.16.1](https://github.com/upb-uc4/University-Credits-4.0/compare/report-v0.16.0...report-v0.16.1) (2021-01-28)
 ## Feature
 ## Refactor